@@ -9,15 +9,24 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "vue": "^3.2.33",
-    "vue-router": "^4.0.14"
+    "@formkit/auto-animate": "^0.8.2",
+    "canvas-confetti": "^1.9.3",
+    "filesize": "^10.1.6",
+    "lucide-vue-next": "^0.446.0",
+    "sticky-js": "^1.3.0",
+    "typed.js": "^2.1.0",
+    "vue": "^3.4.38",
+    "vue-draggable-next": "^2.2.1",
+    "vue-router": "^4.4.3"
   },
   "devDependencies": {
+    "@arco-design/web-vue": "^2.56.0",
     "@vitejs/plugin-vue": "^4.2.3",
     "@vue/compiler-sfc": "^3.2.33",
     "less": "^4.1.2",
     "less-loader": "^10.2.0",
+    "stickybits": "^3.7.11",
     "terser": "^5.19.1",
-    "vite": "^4.4.4"
+    "vite": "^5.4.2"
   }
 }
