@@ -1,4 +1,4 @@
-const modules = import.meta.globEager('./*.vue')
+const modules = import.meta.glob('./*.vue',{eager:true})
 const map = {}
 Object.keys(modules).forEach(file => {
   const modulesName = file.replace('./', '').replace('.vue', '')
